simple_kes/simple_kes_requirements.txt
requests  
torch  
Pyyaml

